refactor: 프로젝트 디렉토리 구조 리팩터링
코드 가독성과 유지보수성을 위해 디렉토리 구조를 재정비했습니다. 주요 변경 사항은 static, templates 등 폴더를 app 내부로 이동하고 import 경로를 수정한 것입니다. README와 테스트 스크립트도 새로운 구조에 맞게 업데이트했습니다.
