+++ app/main.py
@@ -1,10 +1,12 @@
+import uvicorn
 from fastapi import FastAPI, HTTPException, Request
 from fastapi.responses import StreamingResponse, HTMLResponse
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
-from text_to_speech_stream import text_to_speech_stream
-import uvicorn
 from pydantic import BaseModel
+
+from app.service.text_to_speech_stream import text_to_speech_stream
+
 
 # Define request model
 class TextQuery(BaseModel):
@@ -18,10 +20,10 @@ app = FastAPI(
 )
 
 # Mount static files directory
-app.mount("/static", StaticFiles(directory="static"), name="static")
+app.mount("/static", StaticFiles(directory="app/static"), name="static")
 
 # Set up templates
-templates = Jinja2Templates(directory="templates")
+templates = Jinja2Templates(directory="app/templates")
 
 @app.post("/text-to-speech")
 async def convert_text_to_speech(query: TextQuery):
@@ -52,4 +54,4 @@ async def root(request: Request):
 
 # Run the app if this file is executed directly
 if __name__ == "__main__":
-    uvicorn.run("api:app", host="0.0.0.0", port=8000, reload=True)
+    uvicorn.run("app.main:app", host="0.0.0.0", port=8000, reload=True)
