+++ app/templates/index.html
@@ -10,16 +10,16 @@
     <div class="container">
         <h1>Text-to-Speech Converter</h1>
         <p>Enter text below to convert it to speech using ElevenLabs API</p>
-        
+
         <form id="tts-form">
             <div class="form-group">
                 <label for="text-input">Text to convert:</label>
                 <textarea id="text-input" name="text" rows="5" placeholder="Enter your text here..." required></textarea>
             </div>
-            
+
             <button type="submit" id="convert-btn">Convert to Speech</button>
         </form>
-        
+
         <div id="result" class="hidden">
             <h2>Generated Speech</h2>
             <audio id="audio-player" controls></audio>
@@ -27,16 +27,16 @@
                 <button id="download-btn" class="hidden">Download MP3</button>
             </div>
         </div>
-        
+
         <div id="loading" class="hidden">
             <p>Converting text to speech...</p>
         </div>
-        
+
         <div id="error" class="hidden">
             <p>An error occurred. Please try again.</p>
         </div>
     </div>
-    
+
     <script src="/static/js/script.js"></script>
 </body>
 </html>
