+++ app/test/script.py
@@ -1,8 +1,10 @@
 import os
-import time
 import tempfile
+import time
+
 import pygame
 from text_to_speech_stream import text_to_speech_stream
+
 
 def play_audio_stream(audio_stream):
     """
