+++ app/test/test_api.py
@@ -1,5 +1,7 @@
-import requests
 import os
+
+import requests
+
 
 def test_text_to_speech_api():
     """
@@ -32,5 +34,5 @@ def test_text_to_speech_api():
 
 if __name__ == "__main__":
     print("Testing the text-to-speech API...")
-    print("Make sure the API is running (python api.py) before running this test.")
+    print("Make sure the API is running (python app.py) before running this test.")
     test_text_to_speech_api()
